agentic-ai-platform/backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.23.2
motor>=3.3.1
pymongo>=4.6.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.4.2
email-validator>=2.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.16
chromadb>=0.4.22
sentence-transformers>=2.2.2
openai>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
aiohttp>=3.8.5
pytest>=7.4.0
pytest-asyncio>=0.21.1
httpx>=0.25.0